Release (#23)
* feat: Minimize letterbox card when overscrolling (#19)

* fix: iOS release and Second card not overscrolling (#20)

Housekeeping and changes for ios release to function
Reset the nested scroll when new card is rendered.
Also base the offsets on window size instead of card size. more intuitive that way

* feat: Update release script to have appropriate ios version bump steps (#21)

* ci: Trigger android builds on proper tags (#22)

Adds platform to tagging of releases so that the autiomation script can run the appropriate pipeline.

Automation script altered to trigger android releases on correct tags marked as android